refactor to make testing options easier
